+++ src/projects/components/project-detail-page.styles.ts
