+++ src/projects/components/project-modal/project-modal.styles.ts
