+++ src/projects/components/sidebar/projcet-sidebar.styles.ts
